fix(bindings): re-apply C-ABI input validation on the Node and WASM direct-call paths
The Node addon and WASM embind facades call the C++ core directly, bypassing the
C ABI's input validation. Malformed audio (non-finite samples, a sample rate
outside the supported range) or non-finite automation breakpoints therefore
produced garbage output on those surfaces while the Python path, which goes
through the C ABI, rejected the same input.

- Node mastering / effects / repair entries call validate_offline_audio_input
  before handing raw samples to the core, matching the C-ABI guard.
- WASM setAutomationLane rejects non-finite breakpoint ppq/value.
- WASM stream stats totalSamples no longer casts a size_t through a 32-bit int,
  which wrapped negative past 2^31 samples and diverged from Node and Python.

+++ src/wasm/bindings/analysis/stream_analyzer.cpp
@@ -233,7 +233,10 @@ class StreamAnalyzerWrapper {
 
     val out = val::object();
     out.set("totalFrames", s.total_frames);
-    out.set("totalSamples", static_cast<int>(s.total_samples));
+    // A double exactly represents every size_t value that occurs here, and JS
+    // numbers are doubles; static_cast<int> would truncate/wrap streams past
+    // 2^31 samples (~13.5h @44.1kHz) or a reset() with a large base offset.
+    out.set("totalSamples", static_cast<double>(s.total_samples));
     out.set("durationSeconds", s.duration_seconds);
     out.set("pendingFrames", s.pending_frames);
     out.set("droppedOutputFrames", s.dropped_output_frames);
